+++ jslint_ci.sh
@@ -976,6 +976,7 @@ import modulePath from "path";
     let url;
     function httpRequest({
         method,
+        modeSha,
         payload
     }) {
         return new Promise(function (resolve) {
@@ -997,22 +998,35 @@ import modulePath from "path";
                 });
                 res.on("end", function () {
                     responseBuf = Buffer.concat(responseBuf);
-                    moduleAssert.ok(res.statusCode === 200, (
-                        "shGithubFileUpload"
-                        + `- failed to download/upload file ${url} - `
-                        + responseBuf.slice(0, 1024).toString()
-                    ));
+                    moduleAssert.ok(
+                        (
+                            res.statusCode < 400
+                            || (res.statusCode === 404 && modeSha)
+                        ),
+                        (
+                            `shGithubFileUpload - ${res.statusCode}`
+                            + ` - failed to download/upload file ${url} - `
+                            + responseBuf.slice(0, 1024).toString()
+                        )
+                    );
                     resolve();
                 });
             }).end(payload);
         });
     }
+    console.error(
+        content
+        ? `shGithubFileUpload - ${process.argv[1]}`
+        : `shGithubFileDownload - ${process.argv[1]}`
+    );
     path = path.split("/");
     repo = path.slice(0, 2).join("/");
     branch = path[2];
     path = path.slice(3).join("/");
     url = `https://api.github.com/repos/${repo}/contents/${path}`;
-    await httpRequest({});
+    await httpRequest({
+        modeSha: content
+    });
     if (!content) {
         await moduleFs.promises.writeFile(
             modulePath.basename(url),